a criteria is applied; however relative RMS is stuck at 0.055...;

+++ include/models.h
@@ -12,6 +12,6 @@ using Eigen::MatrixXd;
 float dimension(int i, int j);
 float gen();
 void initial_model(std::string elements, MatrixXd Data);
-void least_squares(MatrixXd Data, MatrixXd dims, VectorXd dens, float alpha, float RMS_rel_prev);
+void least_squares(MatrixXd Data, MatrixXd dims, VectorXd dens, float alpha, double RMS_rel_prev);
 
 #endif //MODELS_H
